Created main view and altered active view logic into a method
