+++ hubcare/hubcare_api/hubcare_api/urls.py
@@ -15,9 +15,10 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path
-from hubcare_api.active_views import ActiveQuestion
+# from hubcare_api.active_views import ActiveQuestion
 from hubcare_api.support_views import SupportQuestion
 from hubcare_api.welcoming_views import WelcomingQuestion
+from hubcare_api.views import HubcareApiView
 
 
 support_quest = 'support_question/<str:owner>/<str:repo>/'
@@ -25,8 +26,8 @@ welcoming_quest = 'welcoming_question/<str:owner>/<str:repo>/'
 active_quest = 'active_question/<str:owner>/<str:repo>/'
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path(active_quest, ActiveQuestion.as_view()),
+    # path(active_quest, ActiveQuestion.as_view()),
     path(welcoming_quest, WelcomingQuestion.as_view()),
-    path(support_quest, SupportQuestion.as_view())
-
+    path(support_quest, SupportQuestion.as_view()),
+    path('hubcare_indicators/<str:owner>/<str:repo>/', HubcareApiView.as_view())
 ]
